requirements.txt
PyQt5==5.14.1
pyserial==3.4
matplotlib==3.2.1
numpy==1.18.1
pandas==1.0.3
